started to handle events
