+++ src/AutomatedCar/SystemComponents/InputManager/InputManager.cs
@@ -1,5 +1,6 @@
 ﻿namespace AutomatedCar.SystemComponents.InputManager
 {
+    using AutomatedCar.Models;
     using AutomatedCar.SystemComponents.InputManager.Messenger;
     using AutomatedCar.SystemComponents.Packets;
     using Avalonia.Input;
@@ -12,6 +13,10 @@
         {
             this.InputPacket = new InputPacket();
             virtualFunctionBus.InputPacket = this.InputPacket;
+
+            ControlMessenger.Instance.SteeringEventHandler += OnSteering;
+            ControlMessenger.Instance.PedalEventHandler += OnPedal;
+            ControlMessenger.Instance.GearboxEventHandler += OnGearbox;
         }
 
         private bool IsGearStateJustChanged(Gears newGearState)
@@ -19,52 +24,59 @@
             return this.InputPacket.GearState != newGearState;
         }
 
+        public void OnSteering(object sender, ControlEventArgs eventArgs)
+        {
+            switch (eventArgs.Steering)
+            {
+                case SteeringState.Left:
+                    this.InputPacket.SteeringState = SteeringState.Left;
+                    break;
+                case SteeringState.Right:
+                    this.InputPacket.SteeringState = SteeringState.Right;
+                    break;
+                case SteeringState.Center:
+                    this.InputPacket.SteeringState = SteeringState.Center;
+                    break;
+            }
+        }
+       
+        public void OnPedal(object sender, ControlEventArgs eventArgs)
+        {
+            switch (eventArgs.Pedal)
+            {
+                case Pedals.Gas:
+                    this.InputPacket.PedalState = Pedals.Gas;
+                    break;
+                case Pedals.Empty:
+                    this.InputPacket.PedalState = Pedals.Empty;
+                    break;
+                case Pedals.Brake:
+                    this.InputPacket.PedalState = Pedals.Brake;
+                    break;
+            }
+        }
+
+        public void OnGearbox(object sender, ControlEventArgs eventArgs)
+        {
+            switch (eventArgs.Gear)
+            {
+                case Gears.ShiftUp:
+                    this.InputPacket.GearState = Gears.ShiftUp;
+                    break;
+                case Gears.ShiftDown:
+                    this.InputPacket.GearState = Gears.ShiftDown;
+                    break;
+                case Gears.Steady:
+                    this.InputPacket.GearState = Gears.Steady;
+                    break;
+            }
+        }
+
+        
         public override void Process()
         {
-            // steering
-            if (Keyboard.IsKeyDown(Key.Left))
-            {
-                this.InputPacket.SteeringState = SteeringState.Left;
-            }
-            else if (Keyboard.IsKeyDown(Key.Right))
-            {
-                this.InputPacket.SteeringState = SteeringState.Right;
-            }
-            else
-            {
-                this.InputPacket.SteeringState = SteeringState.Center;
-            }
+            // IsGearStateJustChanged (todo)
 
-            // pedal
-            if (Keyboard.IsKeyDown(Key.Up))
-            {
-                this.InputPacket.PedalState = Pedals.Gas;
-            }
-            else if (Keyboard.IsKeyDown(Key.Down))
-            {
-                this.InputPacket.PedalState = Pedals.Brake;
-            }
-            else
-            {
-                this.InputPacket.PedalState = Pedals.Empty;
-            }
-
-            // gear
-            if (Keyboard.IsKeyDown(Key.PageUp))
-            {
-                this.InputPacket.IsGearStateJustChanged = this.IsGearStateJustChanged(Gears.ShiftUp);
-                this.InputPacket.GearState = Gears.ShiftUp;
-            }
-            else if (Keyboard.IsKeyDown(Key.PageDown))
-            {
-                this.InputPacket.IsGearStateJustChanged = this.IsGearStateJustChanged(Gears.ShiftDown);
-                this.InputPacket.GearState = Gears.ShiftDown;
-            }
-            else
-            {
-                this.InputPacket.GearState = Gears.Steady;
-                this.InputPacket.IsGearStateJustChanged = false;
-            }
         }
     }
 }
